[회원관리/표현] controller 구현

+++ src/api/user/presentation/user.controller.ts
@@ -5,6 +5,7 @@ import {
   Delete,
   Get,
   Param,
+  Patch,
   Post,
   UseInterceptors,
 } from '@nestjs/common';
@@ -12,17 +13,29 @@ import { UserService } from '../application/user.service';
 import {
   CreateUserBody,
   FindOneUserParam,
+  RemoveUserBody,
   RemoveUserResponse,
+  UpdateUserBody,
 } from './user.controller.dto';
 import { UserResponseInterceptor } from '../infrastructure/user.interceptor';
 import { AuthUser } from 'src/api/auth/infrastructure/decorator/auth.decorator';
 import { IAuthResponse } from 'src/api/auth/domain/auth.interface';
+import { Roles } from 'src/api/auth/infrastructure/decorator/role.decorator';
+import { UserRole } from '../domain/user.enum';
+import { Public } from 'src/api/auth/infrastructure/decorator/public.decorator';
+import { AuthService } from 'src/api/auth/application/auth.service';
+import { httpExceptionProvider } from 'src/api/common/provider/exception.provider';
+import { ExceptionMessage } from 'src/api/common/provider/message.provider';
 
 @UseInterceptors(UserResponseInterceptor)
 @Controller('user')
 export class UserController {
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly authService: AuthService,
+  ) {}
 
+  @Roles(UserRole.Admin)
   @Get()
   async findMany(): Promise<IUserResponse[]> {
     const users = await this.userService.findMany();
@@ -37,6 +50,7 @@ export class UserController {
     return user.getDetailResponse();
   }
 
+  @Roles(UserRole.Admin)
   @Get(':user_id')
   async findOne(
     @Param() { user_id: id }: FindOneUserParam,
@@ -45,6 +59,7 @@ export class UserController {
     return user.getResponse();
   }
 
+  @Public()
   @Post()
   async create(@Body() body: CreateUserBody): Promise<IUserDetailResponse> {
     const { username, password, phone, gender, birth_year } = body;
@@ -58,10 +73,22 @@ export class UserController {
     return user.getDetailResponse();
   }
 
-  @Delete(':user_id')
+  @Roles(UserRole.Admin)
+  @Patch()
+  async setRole(@Body() { id, role }: UpdateUserBody): Promise<IUserResponse> {
+    const user = await this.userService.setRole({ id, role });
+    return user.getResponse();
+  }
+
+  @Delete()
   async remove(
-    @Param() { user_id: id }: FindOneUserParam,
+    @AuthUser() auth: IAuthResponse,
+    @Body() { username, password }: RemoveUserBody,
   ): Promise<RemoveUserResponse> {
+    if (auth.username !== username) {
+      throw httpExceptionProvider('403', ExceptionMessage.FBD);
+    }
+    const { id } = await this.authService.validate({ username, password });
     await this.userService.remove({ id });
     return { id };
   }
